demi model

+++ objLoader.hpp
@@ -3,7 +3,7 @@
 
 bool loadAssImp(
 	const char* path,
-	std::vector<unsigned short>& indices,
+	std::vector<unsigned int>& indices,
 	std::vector<glm::vec3>& vertices,
 	std::vector<glm::vec2>& uvs,
 	std::vector<glm::vec3>& normals
